Move more classes into the edu.pdx.cs410J.grader.canvas package.
