+++ grader/src/test/java/edu/pdx/cs410J/grader/canvas/GradesFromCanvasTest.java
@@ -1,5 +1,8 @@
-package edu.pdx.cs410J.grader;
+package edu.pdx.cs410J.grader.canvas;
 
+import edu.pdx.cs410J.grader.Assignment;
+import edu.pdx.cs410J.grader.GradeBook;
+import edu.pdx.cs410J.grader.Student;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
